fix: params to string only on firebase
