+++ src/services/analytics/firebase/index.test.ts
@@ -1,5 +1,10 @@
 import * as CrashReport from "services/crashReport";
-import { logFirebaseEvent, setUserProperties, setUserId } from ".";
+import {
+  logFirebaseEvent,
+  setUserProperties,
+  setUserId,
+  convertParamsToString,
+} from ".";
 
 const mockAnalytics = {
   logEvent: jest.fn(),
@@ -84,5 +89,35 @@ describe("#setUserId", () => {
     const userId = 1;
     setUserId(userId);
     expect(mockAnalytics.setUserId).toHaveBeenCalledWith("1");
+  });
+});
+
+describe("#convertParamsToString", () => {
+  describe("when params are defined", () => {
+    it("converts the params to string", () => {
+      const params = {
+        id: 5,
+        brand: "Brand",
+      };
+
+      const convertedParams = convertParamsToString(params);
+
+      expect(convertedParams.id).toEqual("5");
+      expect(convertedParams.brand).toEqual("Brand");
+    });
+  });
+
+  describe("when there is an undefined param", () => {
+    it("converts the undefined param to an empty string", () => {
+      const params = {
+        id: 5,
+        brand: undefined,
+      };
+
+      const convertedParams = convertParamsToString(params);
+
+      expect(convertedParams.id).toEqual("5");
+      expect(convertedParams.brand).toEqual("");
+    });
   });
 });
